Added Mean Square Model to breeze.interpolation
MeanSquareModel.scala implements a base of functions and a Mean Squares
Model built upon that base. Such model approximate the points used to
build it as alinear combination of the functions in the base.

Be warned, this is an approximator, not an interpolator. This means that
the resulting function is not guaranteed to pass through the sample
points.

Albeith this is not an interpolator, I put it in breeze.interpolation
package as it can be used exactly in the same way of existing
interpolators.